feat : #22 - 툴바 분리작업 완료
