+++ index.html
@@ -15,5 +15,6 @@
   <body>
     <main id="app"></main>
     <script src="/src/main.js" type="module"></script>
+    <div id="toolbar" class="editor__toolbar"></div>
   </body>
 </html>
