+++ src/components/index.js
@@ -1,3 +1,4 @@
 export { default as SideBarHeader } from "./SideBar/SideBarHeader";
 export { default as SideBarPages } from "./SideBar/SideBarPages";
 export { default as EditorTotalContents } from "./Editor/EditorTotalContents";
+export { default as Toolbar } from "./Toolbar";
